// src/models/GameData.ts
export interface Goal {
    team: string
    period: string
    scorer: Scorer
    assists?: Assist[]
    min?: number
    sec?: number
    strength?: string
    emptyNet?: boolean
    showGoal?: boolean
    homeScore?: number;
    awayScore?: number;
    currentGoal?: number;
}

export interface Scorer {
    player: string
    playerId: number
    seasonTotal?: number
}

export interface Assist {
    player: string
    playerId: number
    seasonTotal: number
}